@@ -17,39 +17,39 @@
     "e2e": "ng e2e"
   },
   "peerDependencies": {
-    "@angular/animations": "^7.2.13",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.13",
-    "@angular/compiler": "^7.2.13",
-    "@angular/core": "^7.2.13",
-    "@angular/forms": "^7.2.13",
-    "@angular/http": "^7.2.13",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.13",
-    "@angular/platform-browser-dynamic": "^7.2.13",
-    "@angular/router": "^7.2.13",
-    "@covalent/core": "^2.0.1",
+    "@angular/animations": "^8.0.3",
+    "@angular/cdk": "^8.0.1",
+    "@angular/common": "^8.0.3",
+    "@angular/compiler": "^8.0.3",
+    "@angular/core": "^8.0.3",
+    "@angular/forms": "^8.0.3",
+    "@angular/http": "^7.2.15",
+    "@angular/material": "^8.0.1",
+    "@angular/platform-browser": "^8.0.3",
+    "@angular/platform-browser-dynamic": "^8.0.3",
+    "@angular/router": "^8.0.3",
+    "@covalent/core": "^2.0.2",
     "angular2-text-mask": "^9.0.0",
     "text-mask-addons": "^3.8.0",
     "reflect-metadata": "^0.1.13",
-    "class-transformer": "^0.2.0",
+    "class-transformer": "^0.2.3",
     "chart.js": "^2.8.0",
     "ng2-logger": "^2.0.48",
-    "ngx-device-detector": "^1.3.5",
-    "core-js": "^3.0.1",
-    "rxjs": "^6.4.0",
-    "zone.js": "^0.9.0",
+    "ngx-device-detector": "^1.3.6",
+    "core-js": "^3.1.4",
+    "rxjs": "^6.5.2",
+    "zone.js": "^0.9.1",
     "moment": "2.24.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.9",
-    "@angular/cli": "7.3.9",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "^7.2.15",
-    "@types/jasmine": "~3.3.12",
+    "@angular-devkit/build-angular": "~0.800.6",
+    "@angular/cli": "8.0.6",
+    "@angular/compiler-cli": "^8.0.3",
+    "@angular/language-service": "^8.0.3",
+    "@types/jasmine": "~3.3.13",
     "@types/jasminewd2": "~2.0.6",
-    "@types/node": "~12.0.0",
-    "codelyzer": "~5.0.1",
+    "@types/node": "~12.0.10",
+    "codelyzer": "~5.1.0",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.1.0",
@@ -59,12 +59,12 @@
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.2",
     "moment-es6": "^1.0.0",
-    "ng-packagr": "^5.1.0",
+    "ng-packagr": "^5.3.0",
     "npm-install-peers": "^1.2.1",
     "protractor": "~5.4.2",
-    "ts-node": "~8.1.0",
+    "ts-node": "~8.3.0",
     "tsickle": "^0.35.0",
-    "tslint": "~5.16.0",
-    "typescript": "3.2.4"
+    "tslint": "~5.18.0",
+    "typescript": "3.4.5"
   }
 }
